# docs/swagger.yaml
definitions:
...
  handlers.TaskInput:
    properties:
      description:
        type: string
      priority:
        type: string
      project_id:
        type: integer
      responsible_user_id:
        type: integer
      status:
        type: string
      title:
        type: string
    type: object
...
  models.PriorityEnum:
    enum:
    - low
    - medium
    - high
    type: string
    x-enum-varnames:
    - Low
    - Medium
    - High
...
  models.StatusEnum:
    enum:
    - new
    - in_progress
    - done
    type: string
    x-enum-varnames:
    - New
    - InProgress
    - Done
  models.Task:
    properties:
      completion_date:
        type: string
      creation_date:
        type: string
      description:
        type: string
      id:
        type: integer
      priority:
        $ref: '#/definitions/models.PriorityEnum'
      project_id:
        type: integer
      responsible_user_id:
        type: integer
      status:
        $ref: '#/definitions/models.StatusEnum'
      title:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is project management service API
  title: Project Management Service API
paths:
...
  /tasks:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Task'
            type: array
        "404":
          description: No tasks found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get all tasks
      tags:
      - tasks
    post:
      consumes:
      - application/json
      parameters:
      - description: Task
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/handlers.TaskInput'
      produces:
      - application/json
      responses:
        "201":
          description: Task created
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Create a task
      tags:
      - tasks
  /tasks/{id}:
    delete:
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Task deleted
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "404":
          description: Task not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Delete a task
      tags:
      - tasks
    get:
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Task'
        "404":
          description: Task not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get a task by ID
      tags:
      - tasks
    put:
      consumes:
      - application/json
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Task
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/handlers.TaskInput'
      produces:
      - application/json
      responses:
        "200":
          description: Task updated
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "404":
          description: Task not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Update a task
      tags:
      - tasks
  /tasks/search:
    get:
      parameters:
      - description: Task title
        in: query
        name: title
        type: string
      - description: Task status
        in: query
        name: status
        type: string
      - description: Task priority
        in: query
        name: priority
        type: string
      - description: Task assignee
        in: query
        name: assignee
        type: string
      - description: Task project
        in: query
        name: project
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Task'
            type: array
        "400":
          description: No search parameters provided
          schema:
            type: string
        "404":
          description: No tasks found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Search tasks
      tags:
      - tasks
...
swagger: "2.0"